Adds a reusable `ResolvedRouteValues` instance for all requests where there are no `RouteValues`. 

This reduces allocs and improves perf slightly.
```
|           Method |      Mean |    Error |    StdDev |    Median |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |----------:|---------:|----------:|----------:|-------:|------:|------:|----------:|
|              Old | 222.59 ns | 9.174 ns | 26.616 ns | 212.48 ns | 0.0496 |     - |     - |     208 B |
| OldParameterless |  88.45 ns | 1.761 ns |  2.582 ns |  88.09 ns | 0.0172 |     - |     - |      72 B |
|              New | 202.92 ns | 5.285 ns |  9.117 ns | 201.08 ns | 0.0496 |     - |     - |     208 B |
| NewParameterless |  11.94 ns | 0.273 ns |  0.570 ns |  11.79 ns |      - |     - |     - |         - |
```

Gets rid of the Resolved `ResolvedRouteValues` on `RouteValues` in favour of a lighter `bool` which still supports the `IndexRequest` requirement.